+++ gui/drv_sdl2mw.c
@@ -845,7 +845,7 @@ Edit(void *_Nonnull obj)
 	AG_WindowSetPosition(win, AG_WINDOW_BL, 0);
 
 	lbl = AG_LabelNew(win, 0, _("SDL2MW Driver: %s"), OBJECT(smw)->name);
-	AG_SetFontFamily(lbl, "cm-sans");
+	AG_SetFontFamily(lbl, "charter");
 	AG_SetFontSize(lbl, "150%");
 
 	nb = AG_NotebookNew(win, AG_NOTEBOOK_EXPAND);