+++ gui/joystick.c
@@ -183,7 +183,7 @@ Edit(void *obj)
 	box = AG_BoxNewVert(NULL, AG_BOX_HFILL);
 
 	lbl = AG_LabelNewS(box, 0, AGINPUTDEV(joy)->desc);
-	AG_SetFontFamily(lbl, "cm-sans");
+	AG_SetFontFamily(lbl, "charter");
 	AG_SetFontSize(lbl, "130%");
 
 	AG_LabelNew(box, AG_LABEL_HFILL,